+++ setup.py
@@ -27,7 +27,6 @@ setup(name='distributed',
                 'distributed.bokeh.status',
                 'distributed.bokeh.tasks',
                 'distributed.bokeh.workers',
-                'distributed.bokeh.worker',
                 'distributed.cli',
                 'distributed.deploy',
                 'distributed.diagnostics',
